Add TextWriterLogger
